improve docs; remove `SensorReader`
