+++ Sources/SpeziSensorKit/Sensor/AnySensor.swift
@@ -6,66 +6,40 @@
 // SPDX-License-Identifier: MIT
 //
 
-// swiftlint:disable file_types_order
-
 public import SensorKit
 import SpeziFoundation
 
 
-/// A `Sendable` representation of a data sample obtained from SensorKit.
-public protocol SensorKitSampleSafeRepresentation: Hashable, Sendable {
-    /// The sample's timestamp.
-    ///
-    /// Depending on the specific sensor this sample originates from, the timestamp represents either the actual time the sample was recorded, or the time it was added to SensorKit.
-    var timestamp: Date { get }
-}
-
-
-/// A type that is returned from a SensorKit query.
-///
-/// - Important: For any type that declares conformance to  ``SensorKitSampleProtocol``, the ``SafeRepresentationProcessingInput`` type **must** be equal to the type itself.
-///     Failure to do so is a programmer error and will result in the program crashing at runtime.
-///     This associatedtype is required to work around type system limitations.
-public protocol SensorKitSampleProtocol: AnyObject, Hashable {
-    /// A "safe" `Sendable` representation of the type.
-    ///
-    /// This is required because most SensorKit types are not Sendable and can only be used on the specific DispatchQueue used internally by SensorKit.
-    associatedtype SafeRepresentation: SensorKitSampleSafeRepresentation
-    
-    /// The sample input when processing samples of this type into their safe representation.
-    associatedtype SafeRepresentationProcessingInput: AnyObject, Hashable = Self
-    
-    /// Processes a batch of samples into their safe representation.
-    @inlinable
-    static func processIntoSafeRepresentation(
-        _ samples: some Sequence<(timestamp: Date, sample: SafeRepresentationProcessingInput)>
-    ) throws -> [SafeRepresentation]
-}
-
-
-extension SensorKitSampleProtocol where SafeRepresentation == SafeRepresentationProcessingInput {
-    @inlinable
-    public static func processIntoSafeRepresentation( // swiftlint:disable:this missing_docs
-        _ samples: some Sequence<(timestamp: Date, sample: SafeRepresentationProcessingInput)>
-    ) -> [SafeRepresentation] {
-        samples.map(\.sample)
-    }
-}
-
-extension SensorKitSampleProtocol where SafeRepresentation == DefaultSensorKitSampleSafeRepresentation<SafeRepresentationProcessingInput> {
-    @inlinable
-    public static func processIntoSafeRepresentation( // swiftlint:disable:this missing_docs
-        _ samples: some Sequence<(timestamp: Date, sample: SafeRepresentationProcessingInput)>
-    ) -> [SafeRepresentation] {
-        samples.map { .init(timestamp: $0, sample: $1) }
-    }
-}
-
-
 /// A type-erased ``Sensor``
 ///
-/// - Important: The ``AnySensor`` protocol is public, but your application should not declare any new conformances to it; ``Sensor`` is the only type allowed to conform to ``AnySensor``.
+/// - Important: ``Sensor`` is the only type allowed to conform to ``AnySensor``.
+///
+/// ## Topics
+///
+/// ### Associated types
+/// - ``Sample``
+///
+/// ### Instance Properties
+/// - ``srSensor``
+/// - ``id``
+/// - ``displayName``
+/// - ``authorizationStatus``
+/// - ``dataQuarantineDuration``
+/// - ``suggestedBatchSize``
+///
+/// ### Instance Methods
+/// - ``startRecording()``
+/// - ``stopRecording()``
+/// - ``fetchDevices()``
+/// - ``fetch(from:timeRange:)``
+/// - ``fetch(from:mostRecentAvailable:)``
+///
+/// ### Other
+/// - ``~=(_:_:)``
+/// - ``==(_:_:)-(Sensor<Any>,AnySensor)``
+/// - ``==(_:_:)-(AnySensor,Sensor<Any>)``
 public protocol AnySensor<Sample>: Hashable, Identifiable, Sendable {
+    /// The type of the sensor's resulting samples.
     associatedtype Sample: SensorKitSampleProtocol
     
     /// The underlying SensorKit `SRSensor`.
@@ -77,6 +51,7 @@ public protocol AnySensor<Sample>: Hashable, Identifiable, Sendable {
     /// The sensor's unique identifier.
     var id: String { get }
 }
+
 
 extension AnySensor {
     @inlinable public var id: String { // swiftlint:disable:this missing_docs
@@ -101,6 +76,8 @@ extension AnySensor {
         }
     }
 }
+
+// MARK: Other
 
 /// Compare two sensors, based on their identifiers
 @inlinable
