+++ Sources/SpeziSensorKit/SensorKitFramework+Extensions.swift
@@ -9,6 +9,13 @@
 public import SensorKit
 
 
+extension SRSensorReader {
+    @inlinable
+    convenience init(_ sensor: some AnySensor) {
+        self.init(sensor: sensor.srSensor)
+    }
+}
+
 @available(iOS 17.4, *)
 extension SRElectrocardiogramData.Flags: @retroactive Hashable {
     public func hash(into hasher: inout Hasher) {
